app/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
google-cloud-storage==2.1.0
Jinja2==3.0.3
aiofiles==0.8.0
python-multipart==0.0.5
sendgrid==6.9.7
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pandas==2.0.0